# docs/complete-api-swagger.yaml
openapi: 3.0.0
info:
  title: 🛒 Ghanshyam Murti Bhandar - Complete Ecommerce API
  description: |
    # 🚀 Enterprise-Grade Ecommerce Platform API
    
    **239+ API Endpoints** covering all aspects of modern ecommerce for Android integration.
    
    ## 🔥 Complete Feature Set
    - **🔐 Authentication** - JWT-based secure authentication (4 endpoints)
    - **📦 Products** - Complete product management (12 endpoints)
    - **🏷️ Categories** - Category management (8 endpoints)
    - **🛒 Shopping Cart** - Real-time cart management (6 endpoints)
    - **❤️ Wishlist** - Save favorite products (5 endpoints)
    - **📍 Addresses** - Multi-address support (6 endpoints)
    - **📋 Orders** - Complete order lifecycle (15 endpoints)
    - **💳 Payments** - Razorpay, COD, Wallet (10 endpoints)
    - **🚚 Shipping** - Shiprocket integration (12 endpoints)
    - **💰 Wallet** - Digital wallet system (8 endpoints)
    - **🎫 Coupons** - Discount system (10 endpoints)
    - **⭐ Reviews** - Rating system (8 endpoints)
    - **👥 Users** - User management (10 endpoints)
    - **🏢 Admin** - Admin dashboard (25 endpoints)
    - **📊 Analytics** - Business intelligence (15 endpoints)
    - **📁 File Upload** - Image management (5 endpoints)
    - **🔔 Notifications** - Real-time updates (6 endpoints)
    - **🧾 Invoices** - GST compliance (8 endpoints)
    - **📦 Inventory** - Stock management (12 endpoints)
    - **🏪 Suppliers** - Vendor management (10 endpoints)
    
    ## 🎯 Android Integration Ready
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token for authentication. Get token from `/auth/login`.
        
        **Usage:** `Authorization: Bearer <your-jwt-token>`
        
        **Example:** `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Request success status
          example: true
        message:
          type: string
          description: Response message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: "2025-07-25T14:30:00Z"
      required:
        - success
        - message

    # Error Response
...
paths:
...
  /orders:
    get:
      tags:
        - 📋 Orders
      summary: 📦 Get User Orders
      description: |
        Retrieve current user's orders with pagination and filtering.

        **Android Integration:**
        - Use for order history screen
        - Implement pull-to-refresh
        - Show order status with colors
        - Enable order tracking
      operationId: getUserOrders
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number
          example: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
          description: Orders per page
          example: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, shipped, delivered, cancelled, returned]
          description: Filter by order status
          example: "shipped"
      responses:
        '200':
          description: ✅ Orders retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          orders:
                            type: array
                            items:
                              type: object
                              properties:
                                _id:
                                  type: string
                                  example: "60f1b2b3c4d5e6f7g8h9i0j6"
                                orderNumber:
                                  type: string
                                  example: "ORD-2025-001"
                                status:
                                  type: string
                                  enum: [pending, confirmed, shipped, delivered, cancelled]
                                  example: "shipped"
                                items:
                                  type: array
                                  items:
                                    type: object
                                    properties:
                                      product:
                                        type: object
                                        properties:
                                          _id:
                                            type: string
                                          name:
                                            type: string
                                          images:
                                            type: array
                                            items:
                                              type: string
                                      quantity:
                                        type: integer
                                      price:
                                        type: number
                                pricing:
                                  type: object
                                  properties:
                                    total:
                                      type: number
                                      example: 2500
                                paymentInfo:
                                  type: object
                                  properties:
                                    method:
                                      type: string
                                      enum: [cod, razorpay, wallet]
                                      example: "razorpay"
                                    status:
                                      type: string
                                      enum: [pending, completed, failed]
                                      example: "completed"
                                createdAt:
                                  type: string
                                  format: date-time
                                  example: "2025-07-25T10:30:00Z"
                          pagination:
                            type: object
                            properties:
                              currentPage:
                                type: integer
                              totalPages:
                                type: integer
                              totalItems:
                                type: integer
                              hasNext:
                                type: boolean
                              hasPrev:
                                type: boolean
        '401':
          description: ❌ Unauthorized
        '500':
          description: ❌ Internal server error

    post:
      tags:
        - 📋 Orders
      summary: 🛒 Create Order
      description: |
        Create a new order from cart items with payment processing.

        **Android Integration:**
        - Use for checkout process
        - Handle payment method selection
        - Show order confirmation
        - Redirect to payment gateway if needed

        **Payment Methods:**
        - **COD**: Cash on Delivery (no payment processing)
        - **Razorpay**: Online payment (returns payment details)
        - **Wallet**: Use wallet balance
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - addressId
                - paymentInfo
              properties:
                addressId:
                  type: string
                  description: Shipping address ID
                  example: "60f1b2b3c4d5e6f7g8h9i0j7"
                paymentInfo:
                  type: object
                  required:
                    - method
                  properties:
                    method:
                      type: string
                      enum: [cod, razorpay, wallet]
                      example: "razorpay"
                      description: Payment method
                couponCode:
                  type: string
                  description: Discount coupon code
                  example: "WELCOME10"
                notes:
                  type: string
                  description: Order notes
                  example: "Please handle with care"
            examples:
              cod_order:
                summary: COD Order
                value:
                  addressId: "60f1b2b3c4d5e6f7g8h9i0j7"
                  paymentInfo:
                    method: "cod"
                  notes: "Cash on delivery order"
              razorpay_order:
                summary: Razorpay Order
                value:
                  addressId: "60f1b2b3c4d5e6f7g8h9i0j7"
                  paymentInfo:
                    method: "razorpay"
                  couponCode: "WELCOME10"
      responses:
        '201':
          description: ✅ Order created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          order:
                            type: object
                            properties:
                              _id:
                                type: string
                              orderNumber:
                                type: string
                              status:
                                type: string
                              pricing:
                                type: object
                                properties:
                                  total:
                                    type: number
                          razorpay:
                            type: object
                            description: Razorpay order details (only for razorpay payment)
                            properties:
                              orderId:
                                type: string
                                example: "order_QxLBGcxGWpimfg"
                              amount:
                                type: integer
                                example: 250000
                                description: Amount in paise
                              currency:
                                type: string
                                example: "INR"
                              keyId:
                                type: string
                                example: "rzp_test_4hUj1dxGbUR5wj"
                          requiresPayment:
                            type: boolean
                            example: true
                            description: Whether payment processing is required
        '400':
          description: ❌ Invalid request or insufficient stock
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Address not found or empty cart
        '500':
          description: ❌ Internal server error
  /orders/{id}:
    get:
      tags:
        - 📋 Orders
      summary: 🔍 Get Order Details
      description: |
        Get detailed information about a specific order.

        **Android Integration:**
        - Use for order detail screen
        - Show order tracking
        - Display payment status
        - Enable order actions (cancel, return)
      operationId: getOrderById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Order ID
          example: "60f1b2b3c4d5e6f7g8h9i0j6"
      responses:
        '200':
          description: ✅ Order details retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          order:
                            type: object
                            properties:
                              _id:
                                type: string
                              orderNumber:
                                type: string
                              status:
                                type: string
                              items:
                                type: array
                                items:
                                  type: object
                              pricing:
                                type: object
                              paymentInfo:
                                type: object
                              shippingAddress:
                                type: object
                              tracking:
                                type: object
                                properties:
                                  awbCode:
                                    type: string
                                  courierName:
                                    type: string
                                  trackingUrl:
                                    type: string
                              timeline:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    status:
                                      type: string
                                    timestamp:
                                      type: string
                                      format: date-time
                                    description:
                                      type: string
        '401':
          description: ❌ Unauthorized
        '404':
          description: ❌ Order not found
        '500':
          description: ❌ Internal server error

  # 💳 PAYMENT ENDPOINTS
...